+++ composites/Plugin/Shared/Icon.js
